fix: change file structure to fix type checking export
