+++ src/app/api/applications/[id]/save-forms-and-duos/route.ts
@@ -2,8 +2,8 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
 import { ExtendedSession } from '@/app/api/auth/auth.types';
+import { authOptions } from '@/app/api/auth/config';
 import { handleErrorResponse } from '@/app/api/errorHandling';
 import { saveFormAndDuos } from '@/services/daam/index.server';
 import { getServerSession } from 'next-auth';
