+++ src/app/api/applications/route.ts
@@ -5,9 +5,9 @@
 import { createApplication, listApplications } from '@/services/daam/index.server';
 import { getServerSession } from 'next-auth';
 import { NextResponse } from 'next/server';
-import { authOptions } from '../auth/[...nextauth]/route';
 import { ExtendedSession } from '../auth/auth.types';
 import { handleErrorResponse } from '../errorHandling';
+import { authOptions } from '../auth/config';
 
 export async function POST(request: Request) {
   const session: ExtendedSession | null = await getServerSession(authOptions);
