+++ src/app/api/auth/auth.ts
@@ -6,8 +6,8 @@ import { decrypt } from '@/utils/encryption';
 import { jwtDecode } from 'jwt-decode';
 import { Account, getServerSession } from 'next-auth';
 import type { JWT } from 'next-auth/jwt';
-import { authOptions } from './[...nextauth]/route';
 import { ExtendedSession } from './auth.types';
+import { authOptions } from './config';
 
 export async function getToken(tokenType: 'access_token' | 'id_token') {
   const session = (await getServerSession(authOptions)) as ExtendedSession;
