+++ src/app/api/auth/logout/route.ts
@@ -3,8 +3,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { getServerSession } from 'next-auth';
-import { authOptions } from '../[...nextauth]/route';
 import { getToken } from '../auth';
+import { authOptions } from '../config';
 
 export async function GET() {
   const session = await getServerSession(authOptions);
